refactor!: update project structure
