--- tap_core/src/adapters/mock.rs
@@ -1,4 +1,0 @@
-// Copyright 2023-, Semiotic AI, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod executor_mock;
