--- tap_core/src/adapters/test/mod.rs
@@ -1,6 +1,0 @@
-// Copyright 2023-, Semiotic AI, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod escrow_adapter_test;
-pub mod rav_storage_adapter_test;
-pub mod receipt_storage_adapter_test;
