+++ tap_core/src/rav/request.rs
@@ -3,10 +3,9 @@
 
 use serde::{Deserialize, Serialize};
 
-use super::{SignedRAV, SignedReceipt};
 use crate::{
-    receipt_aggregate_voucher::ReceiptAggregateVoucher,
-    tap_receipt::{Failed, ReceiptWithState},
+    rav::{ReceiptAggregateVoucher, SignedRAV},
+    receipt::{Failed, ReceiptWithState, SignedReceipt},
 };
 
 #[derive(Debug, Serialize, Deserialize)]
