+++ tap_core/src/signed_message.rs
