+++ tap_integration_tests/tests/lib.rs
@@ -1,5 +1,5 @@
 // Copyright 2023-, Semiotic AI, Inc.
 // SPDX-License-Identifier: Apache-2.0
-
-mod indexer_mock;
-mod showcase;
+//
+// mod indexer_mock;
+// mod showcase;
